% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mylogit.predict.R
\name{mylogit.predict}
\alias{mylogit.predict}
\title{mylogit.predict}
\usage{
mylogit.predict(model, data = NULL, print = FALSE)
}
\arguments{
\item{model}{a logistic model fitted by mylogit function.}

\item{data}{a data frame containing the variables in the model.}

\item{print}{TRUE, print out prediction class; FALSE, as default, no printing.}
}
\value{
a list containing the following components:

\item{yhat}{fitted values}
}
\description{
mylogit.predict is used to make a class prediction of logistic model.
}
\examples{
logit = mylogit(formula = supp ~ len + dose, data = ToothGrowth)
logit.pred = mylogit.predict(logit)

}
