% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mylogit.R
\name{mylogit}
\alias{mylogit}
\title{mylogit}
\usage{
mylogit(formula, data, format = "brief")
}
\arguments{
\item{formula}{a symbolic description of the model to be fitted.}

\item{data}{a data frame containing the variables in the model.}

\item{format}{output format. \emph{brief} for brief output, \emph{detailed} for detailed output.}
}
\value{
a list containing the following components:

\item{coefficients}{coefficients of logistic regression model}

\item{fitted.values}{fitted values for origin dataset}

\item{cov}{covariance matrix}

\item{std.err}{standard error}

\item{deviance}{deviance}

\item{null.deviance}{null deviance}

\item{r2}{r-square, 1 - (deviance/null.deviance)}
}
\description{
mylogit is used to fit a logistic regression model.
}
\examples{
logit = mylogit(formula = supp ~ len + dose, data = ToothGrowth)

}
