% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mylogit.R
\name{mylogit}
\alias{mylogit}
\title{mylogit}
\usage{
mylogit(formula, data, format = "brief")
}
\arguments{
\item{formula}{a symbolic description of the model to be fitted.}

\item{data}{a data frame containing the variables in the model.}

\item{format}{output format. `brief` for brief output, `detailed` for detailed output.}
}
\value{
a list containing the following components:
}
\description{
mylogit is used to fit a logistic regression model.
}
\examples{
mylogit(formula = supp ~ len + dose, data = ToothGrowth)

}
