% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mylogit.R
\name{mylogit}
\alias{mylogit}
\title{mylogit}
\usage{
mylogit(formula, data, format = "brief")
}
\arguments{
\item{formula}{a symbolic description of the model to be fitted.}

\item{data}{a data frame containing the variables in the model.}

\item{format}{output format. \emph{brief}, as default for brief output,
\emph{detailed} for detailed output,
\emph{nothing} for no printing output.}
}
\value{
a list containing the following components:

\item{coefficients}{coefficients of logistic regression model}

\item{fitted.values}{fitted values for input dataset}

\item{cov}{covariance matrix}

\item{std.err}{standard error}

\item{deviance}{deviance}

\item{null.deviance}{null deviance}

\item{r2}{r-square, 1 - (deviance/null.deviance)}

\item{AIC}{AIC score}

\item{null.df}{degrees of freedom of null model}

\item{df}{degrees of freedom of logistic model}

\item{level}{the levels of response variable}

\item{predict.class}{class prediction for input dataset}

\item{predict.accuracy}{accuracy of class prediction}

\item{formula}{formula of logistic model}

\item{data}{input dataset}
}
\description{
mylogit is used to fit a logistic regression model.
}
\examples{
mylogit(formula = supp ~ len + dose, data = ToothGrowth, format="brief")

}
